zephyr: kpb: fix a compiler warning
When built under Zephyr the wrong format generates a compiler
warning:

modules/audio/sof/zephyr/include/sof/trace/trace.h:35:11: warning: \
format '%d' expects argument of type 'int', but argument 3 has \
type 'uint64_t'

Since native SOF tracing cannot print 64-bit data, only print the
low 32 bits of the time interval.

+++ src/audio/kpb.c
@@ -1119,9 +1119,9 @@ static void kpb_init_draining(struct comp_dev *dev, struct kpb_client *cli)
 							   host_period_size / bytes_per_ms) /
 					 KPB_DRAIN_NUM_OF_PPL_PERIODS_AT_ONCE;
 			period_bytes_limit = host_period_size;
-			comp_info(dev, "kpb_init_draining(): sync_draining_mode selected with interval %d [uS].",
-				  drain_interval * 1000 /
-				  clock_ms_to_ticks(PLATFORM_DEFAULT_CLOCK, 1));
+			comp_info(dev, "kpb_init_draining(): sync_draining_mode selected with interval %u [uS].",
+				  (unsigned int)(drain_interval * 1000 /
+						 clock_ms_to_ticks(PLATFORM_DEFAULT_CLOCK, 1)));
 		} else {
 			/* Unlimited draining */
 			drain_interval = 0;
